Everything is done
created a QuiznessAccountManager that adds a grain of salt to the
unencrypted password, hashes that function and stores it.  Returns
booleans with successful logins and successful account creations.
However, I could not test it because for some reason I am getting an
"Error: Could not find or load main class
control.QuiznessAccountManager".

For some reason my LoginServlet and CreateNewAccount also are not
running so testing has not been done.